Better readability, removed some non-sense
